+++ app/src-ui/src/App.tsx
@@ -1,6 +1,7 @@
 import { HashRouter, Routes, Route, NavLink } from "react-router-dom";
 import { SettingsPage } from "./pages/SettingsPage";
 import { SyncPage } from "./pages/SyncPage";
+import { PuzzlePage } from "./pages/PuzzlePage";
 import "./App.css";
 
 function App() {
@@ -13,12 +14,14 @@ function App() {
             <NavLink to="/" end>
               Sync
             </NavLink>
+            <NavLink to="/puzzles">Puzzles</NavLink>
             <NavLink to="/settings">Settings</NavLink>
           </div>
         </nav>
         <main className="app-main">
           <Routes>
             <Route path="/" element={<SyncPage />} />
+            <Route path="/puzzles" element={<PuzzlePage />} />
             <Route path="/settings" element={<SettingsPage />} />
           </Routes>
         </main>